Move generate_patch_url function to utils
