+++ vulnerabilities/improvers/__init__.py
@@ -30,10 +30,10 @@ from vulnerabilities.pipelines.v2_improvers import enhance_with_kev as enhance_w
 from vulnerabilities.pipelines.v2_improvers import (
     enhance_with_metasploit as enhance_with_metasploit_v2,
 )
+from vulnerabilities.pipelines.v2_improvers import fetch_patch_url as fetch_patch_url_v2
 from vulnerabilities.pipelines.v2_improvers import flag_ghost_packages as flag_ghost_packages_v2
 from vulnerabilities.pipelines.v2_improvers import relate_severities
 from vulnerabilities.pipelines.v2_improvers import unfurl_version_range as unfurl_version_range_v2
-from vulnerabilities.pipelines.v2_improvers import fetch_patch_url as fetch_patch_url_v2
 from vulnerabilities.utils import create_registry
 
 IMPROVERS_REGISTRY = create_registry(
